+++ src/main/kotlin/lotto/domain/LottoStrategy.kt
@@ -1,0 +1,5 @@
+package lotto.domain
+
+interface LottoStrategy {
+    fun generateLotto(quantity: Int): Lottoes
+}
